fix: Team 관련 API 호출(레포지토리) 분류
